Refactoring endpoints with express router, routes and controllers
