+++ server/middleware.js
@@ -21,6 +21,3 @@ module.exports.isTeacher = (req, res, next) => {
     };
     next();
 }
-
-
-//! Add TryCatch middleware for async functions
