+++ server/utils.js
@@ -1,0 +1,5 @@
+module.exports.TryCatch = func => {
+    return (req, res, next) => {
+        func(req, res, next).catch(next);
+    }
+}
